REFRACTORED unit test code for passwordless
